api/requirements.txt
fastapi
uvicorn
orjson
openai
python-dotenv